+++ R/textSnippets.R
@@ -295,3 +295,28 @@ inputText <- function(expTypeLevel) {
                "object. ")
     }
 }
+
+#' @rdname textSnippets
+#' @export
+featureCollectionText <- function(featureCollections) {
+    if (length(featureCollections) == 0) {
+        out <- paste0("No feature collections were tested.")
+    } else {
+        out <- ""
+        if ("complexes" %in% featureCollections) {
+            out <- paste0(out, "Complexes are obtained from Corum, CYC2008, ",
+                          "PomBase, HuMAP2 and the Complex Portal (see ",
+                          "the `makeComplexDB` function for more details. ")
+        }
+        if ("GO" %in% featureCollections) {
+            out <- paste0(out, "GO terms are obtained from MSigDB. ")
+        }
+        if ("pathways" %in% featureCollections) {
+            out <- paste0(out, "Pathway information is obtained from ",
+                          "BIOCARTA, KEGG, PID, REACTOME and WIKIPATHWAYS, ",
+                          "via MSigDB (see the `prepareFeatureCollections` ",
+                          "function for more details). ")
+        }
+    }
+    out
+}